--- a/scripts/test-integration.sh
+++ b/scripts/test-integration.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+# 統合テスト（GUI含む）を実行
+
+echo "Running integration tests..."
+npx jest src/__tests__/App.test.tsx --verbose