--- a/scripts/test-unit.sh
+++ b/scripts/test-unit.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+# 単体テストのみを高速実行
+
+echo "Running unit tests only..."
+npx jest src/__tests__/argument-parser.test.ts --verbose