--- a/src/__tests__/App.test.tsx
+++ b/src/__tests__/App.test.tsx
@@ -1,9 +1,17 @@
-import '@testing-library/jest-dom';
 import { render } from '@testing-library/react';
 import App from '../renderer/App';
 
 describe('App', () => {
-  it('should render', () => {
-    expect(render(<App />)).toBeTruthy();
+  it('should render without crashing', () => {
+    const { container } = render(<App />);
+    // 基本的なレンダリングテストのみ
+    expect(container).toBeTruthy();
+  });
+
+  it('should render file list section', () => {
+    const { container } = render(<App />);
+    // ファイルリストセクションが存在することを確認
+    const fileListSection = container.querySelector('.file-list');
+    expect(fileListSection).toBeTruthy();
   });
 });