## Summary
- Optimize test performance - reduce execution time by 65% (70s → 20s)
- Rename SPEC.md to CLAUDE.md for better Claude Code integration
- Improve project documentation and task management

## Test Performance Improvements
- **Single test optimization**: Add comprehensive mocking for Electron API and react-syntax-highlighter
- **Test separation**: Create separate unit tests (test:unit, ~11s) and integration tests (test:integration, ~13s)  
- **Mock infrastructure**: Create setup.ts for shared test configuration and mocking
- **Performance gains**:
  - argument-parser.test.ts: 14s → 11s
  - App.test.tsx: 48s → 13s (heavy mocking of GUI components)
  - **Total: 70s → 20s (65% reduction)**

## Documentation Improvements
- **CLAUDE.md**: Rename SPEC.md to CLAUDE.md for better Claude Code handling
- **Workflow optimization**: Update development workflow for conditional test execution
- **Task management**: Eliminate duplicate management between SPEC.md and TODO.md
- **Process documentation**: Document optimized test commands and workflow

## Test plan
- [x] Unit tests pass in ~11 seconds
- [x] Integration tests pass in ~13 seconds  
- [x] All tests combined pass in ~20 seconds
- [x] Development workflow updated in documentation
- [x] TODO tracking accurately reflects completed features

🤖 Generated with [Claude Code](https://claude.ai/code)